Indexing done.
Different: now we get indexes from fbx vertex normal and uv list. And
later make vertexes with those indexes.
This should make the importer more flexible in the future.
Also now code removes copies of vertexes. And in draw calls the program
instead uses the earliest version of that vertex.